+++ src/components/preset/templates/MemoryTemplate.tsx
@@ -1,8 +1,14 @@
 
 import React from 'react';
+import ActualMemoryTemplate from '../../quiz/preset-games/templates/MemoryTemplate';
 
-const MemoryTemplate = () => {
-  return <div>Memory Template Placeholder</div>;
+interface MemoryTemplateProps {
+  content: any;
+  topic: string;
+}
+
+const MemoryTemplate: React.FC<MemoryTemplateProps> = ({ content, topic }) => {
+  return <ActualMemoryTemplate content={content} topic={topic} />;
 };
 
 export default MemoryTemplate;