+++ src/components/preset/templates/OrderingTemplate.tsx
@@ -1,8 +1,14 @@
 
 import React from 'react';
+import ActualOrderingTemplate from '../../quiz/preset-games/templates/OrderingTemplate';
 
-const OrderingTemplate = () => {
-  return <div>Ordering Template Placeholder</div>;
+interface OrderingTemplateProps {
+  content: any;
+  topic: string;
+}
+
+const OrderingTemplate: React.FC<OrderingTemplateProps> = ({ content, topic }) => {
+  return <ActualOrderingTemplate content={content} topic={topic} />;
 };
 
 export default OrderingTemplate;