+++ src/components/preset/templates/PictionaryTemplate.tsx
@@ -1,8 +1,14 @@
 
 import React from 'react';
+import ActualPictionaryTemplate from '../../quiz/preset-games/templates/PictionaryTemplate';
 
-const PictionaryTemplate = () => {
-  return <div>Pictionary Template Placeholder</div>;
+interface PictionaryTemplateProps {
+  content: any;
+  topic: string;
+}
+
+const PictionaryTemplate: React.FC<PictionaryTemplateProps> = ({ content, topic }) => {
+  return <ActualPictionaryTemplate content={content} topic={topic} />;
 };
 
 export default PictionaryTemplate;