+++ src/components/preset/templates/WordSearchTemplate.tsx
@@ -1,8 +1,14 @@
 
 import React from 'react';
+import ActualWordSearchTemplate from '../../quiz/preset-games/templates/WordSearchTemplate';
 
-const WordSearchTemplate = () => {
-  return <div>WordSearch Template Placeholder</div>;
+interface WordSearchTemplateProps {
+  content: any;
+  topic: string;
+}
+
+const WordSearchTemplate: React.FC<WordSearchTemplateProps> = ({ content, topic }) => {
+  return <ActualWordSearchTemplate content={content} topic={topic} />;
 };
 
 export default WordSearchTemplate;